network: Fix ListErrorModel since UIDs are uint64_t values

+++ examples/error-model/simple-error-model.cc
@@ -158,7 +158,7 @@ main(int argc, char* argv[])
 
     // Now, let's use the ListErrorModel and explicitly force a loss
     // of the packets with pkt-uids = 11 and 17 on node 2, device 0
-    std::list<uint32_t> sampleList;
+    std::list<uint64_t> sampleList;
     sampleList.push_back(11);
     sampleList.push_back(17);
     // This time, we'll explicitly create the error model we want
